This closes #42.

+++ libs/json/src/json_encode.c
@@ -62,13 +62,13 @@ json_encode_value(struct json_encoder *encoder, struct json_value *jv)
             encoder->je_write(encoder->je_arg, encoder->je_encode_buf, len);
             break;
         case JSON_VALUE_TYPE_UINT64:
-            len = sprintf(encoder->je_encode_buf, "%lu",
-                    (unsigned long)jv->jv_val.u);
+            len = sprintf(encoder->je_encode_buf, "%llu",
+                    jv->jv_val.u);
             encoder->je_write(encoder->je_arg, encoder->je_encode_buf, len);
             break;
         case JSON_VALUE_TYPE_INT64:
-            len = sprintf(encoder->je_encode_buf, "%ld",
-                    (long) jv->jv_val.u);
+            len = sprintf(encoder->je_encode_buf, "%lld",
+                    jv->jv_val.u);
             encoder->je_write(encoder->je_arg, encoder->je_encode_buf, len);
             break;
         case JSON_VALUE_TYPE_STRING:
